rename the Driver/Channel combo into DS402Driver/DS402Channel
Either I'm too stupid to understand it, or the Roboteq
implementation of the DS402 profile is totally broken. Either way,
it's time to spend less time debugging this.

Rename both to reflect their usage of the DS402 profile, to pave
the way to start a plain CANOpen controller using the non-DS402
objects.