assume a linear scale between min and max for position, velocity and current
The current code was very badly trying to implement a bilinear approach,
where [min, 0] and [0, max] use different slopes.

To say the least, it was bad and buggy. Also, unused.

So, get rid of it for now. We can always reintroduce the functionality
in case it proves to be useful